commit-1 14/12/20

+++ Customer Site/include/complain.php
@@ -1,7 +1,7 @@
 <?php
         require_once "pdo.php";
 
-        if (isset($_POST['confirm'])) {
+        if (isset($_POST['confirm']) && $_POST['confirm']==='YES') {
 
                 $delete = $pdo->prepare("DELETE FROM inquiry WHERE  Connection_Number = ".$_SESSION['ID_number'].";");
                 $delete->execute();
@@ -9,7 +9,7 @@
                 $data = [
                         ':value1' => '1',
                         ':id' => $_SESSION['ID_number'],
-                    ];
+                ];
 
                 $sql = "UPDATE customer SET Connection_Error = :value1 WHERE ID=:id";
                 $pdo->prepare($sql)->execute($data);
@@ -24,6 +24,7 @@
                 );
                 header("Location: home.php");
                 return;
+                
 
         }
